Schedule entry delay type is now float
Since the rest of schema and classes use floating point numbers,
it made sense to change this for the sake of consistency, however
it may break scripts using the dictionary to access the schedule.

+++ ldfparser/schedule.py
@@ -13,7 +13,7 @@ class ScheduleTable():
 class ScheduleTableEntry():
 
     def __init__(self) -> None:
-        self.delay: int = 0
+        self.delay: float = 0.0
 
 class LinFrameEntry(ScheduleTableEntry):
